+++ core/src/com/colegiovivas/lapandemia/levels/Fan.java
@@ -1,10 +1,10 @@
 package com.colegiovivas.lapandemia.levels;
 
 public class Fan {
-    public float x;
-    public float y;
+    public int x;
+    public int y;
 
-    public Fan(float x, float y) {
+    public Fan(int x, int y) {
         this.x = x;
         this.y = y;
     }
